split out code
